charlies-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9 ; platform_system != "Windows"
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
feedparser==6.0.11
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
APScheduler==3.10.4
python-slugify==8.0.4

